requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
chromadb
sentence-transformers
requests
python-dotenv
genai